requirements.txt
discord.py
requests
pyinstaller